+++ packages/react-components/react-teaching-popover-preview/src/components/TeachingPopoverTitle/useTeachingPopoverTitleStyles.styles.ts
@@ -60,7 +60,7 @@ export const useTeachingPopoverTitleStyles_unstable = (state: TeachingPopoverTit
     state.root.className,
   );
 
-  if (state.showDismiss && state.dismissButton) {
+  if (state.dismissButton) {
     state.dismissButton.className = mergeClasses(
       teachingPopoverTitleClassNames.dismissButton,
       styles.dismissButton,